# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Aegis Platform API
  version: v1
  description: >
    HTTP API for the Aegis enterprise access-control platform — a multi-tenant,
    microservices SaaS substrate (centralized RBAC + ABAC PDP/PEP, PostgreSQL
    row-level tenant isolation, signed service-to-service calls, hash-chained
    audit). Seven business services sit behind a single gateway.


    **Edge routing.** All authenticated traffic enters through the gateway, which
    resolves the first path segment (`user-management`, `expense`, `payroll`,
    `reporting`, `workflow`, `notification`, `invoice`) to a downstream service
    and forwards the request unchanged. Each service independently re-validates
    auth via its own PEP (defense in depth). Every external path therefore has the
    form `/{service}/v1/...` (public prefix `/v1`). `/health` on each service and
    on the gateway bypasses auth/context middleware.


    **Required context headers.** Authenticated requests must carry
    `X-Tenant-Id` and `X-Correlation-Id` (the gateway mints the correlation id at
    the edge if absent and echoes it on every response, including error hops).
    `X-Caller` is set to `gateway` on every forwarded hop. The tenant is ambient
    (driven into PostgreSQL RLS via `SET LOCAL app.current_tenant`) — there is no
    tenant path segment. Money/state-changing writes may honor an
    `Idempotency-Key` header (pay-run disburse).


...
    the principal also holds `payroll.sensitive.read`.


...
  contact:
...
servers:
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Authenticated user JWT issued by user-management `/auth/login`. Carried as
        `Authorization: Bearer <token>`. The gateway propagates it downstream; each
        service validates it and resolves the principal's permissions.
  parameters:
    TenantHeader:
      name: X-Tenant-Id
      in: header
      required: true
      description: Tenant (organization) the request operates within. Drives PostgreSQL RLS.
      schema:
        type: string
    CorrelationHeader:
      name: X-Correlation-Id
      in: header
      required: false
      description: >
        Request-tracking id. Minted at the gateway if absent and echoed on every
        response (including error hops). Required on every internal hop.
      schema:
        type: string
        format: uuid
...
  schemas:
    ErrorEnvelope:
      type: object
      description: Terminal error shape emitted by the service-core error middleware (and the gateway).
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
                description: Stable machine code, e.g. E_VALIDATION, E_UNAUTHORIZED, E_FORBIDDEN, E_NOT_FOUND, E_CONFLICT, E_GATEWAY_TIMEOUT.
              type:
                type: string
                description: Error category, e.g. VALIDATION, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, GATEWAY, SYSTEM.
              message:
                type: string
              correlationId:
                type: string
            required: [code, type, message, correlationId]
      required: [errors]
    DataEnvelope:
      type: object
      description: Single-entity envelope.
      properties:
        data:
          type: object
      required: [data]
...
    ApproveRequest:
      type: object
      properties:
        comment: { type: string }
...
    ApproveInvoiceRequest:
      type: object
      properties:
        comment: { type: string }
        approvalLevel: { type: integer, minimum: 1 }
...
  responses:
    BadRequest:
      description: Validation error (E_VALIDATION).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: Missing/invalid token or required headers (E_UNAUTHORIZED).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Forbidden:
      description: Authenticated but lacks the required permission, or ABAC policy denied (E_FORBIDDEN).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: Resource not found or RLS-invisible (E_NOT_FOUND).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Conflict:
      description: Illegal state transition or uniqueness conflict (E_CONFLICT).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    ServerError:
      description: Unexpected server error (E_SYSTEM / E_DATABASE).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
paths:
...
  /expense/v1/reports/{id}/approve:
    post:
      tags: [expense]
      summary: Approve a report (alias for decisions{approved})
      description: >
        Backward-compatible alias for POST /reports/{id}/decisions
        {decision:'approved'}. Requires permission `expense.report.approve`
        (ABAC amount-cap applies).
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/CorrelationHeader'
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApproveRequest' }
      responses:
        '200':
          description: Report approved.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DataEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/ServerError' }
...
  /expense/v1/reports/{id}:
    get:
      tags: [expense]
      summary: Get a single report by id
      description: Requires permission `expense.report.view` (ABAC resource = report). 404 if RLS-invisible.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/CorrelationHeader'
      responses:
        '200':
          description: Report.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DataEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }

  # ===================== INVOICE =====================
...
  /invoice/v1/invoices/{id}:
    get:
      tags: [invoice]
      summary: Get a single invoice by id
      description: Requires permission `invoice.view`.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/CorrelationHeader'
      responses:
        '200':
          description: Invoice.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DataEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
...
  /invoice/v1/invoices/{id}/approve:
    post:
      tags: [invoice]
      summary: Approve an invoice (alias for decisions{approved})
      description: Backward-compatible alias. Requires permission `invoice.approve`.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/CorrelationHeader'
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApproveInvoiceRequest' }
      responses:
        '200':
          description: Invoice approved.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DataEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/ServerError' }

  # ===================== PAYROLL =====================
...